requirements.txt
discord
pynacl
pafy
youtube_dl
python-dotenv